backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
openai
python-dotenv
fastapi-cors
